requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
backoff==2.1.2
charset-normalizer==2.1.0
frozenlist==1.3.1
gql==3.4.0
graphql-core==3.2.1
multidict==6.0.2
requests==2.28.1
PyYAML~=5.3.1